requirements.txt
pandas>=2.0
streamlit>=1.30
openpyxl>=3.1
plotly>=5.18
numpy>=1.24
psycopg2-binary>=2.9
